Calling np.dtype on a delayed object works (#4387)
* Calling np.dtype on a delayed object works

Calling `np.dtype(dask.delayed(...))` used to result in a segfault, as
numpy recursively tries to get `dtype` from the object. This is likely a
bug in numpy. For now, we can do a dumb for if `x.dtype().dtype()` is
called (which shouldn't ever show up in real code).

See https://github.com/dask/dask/pull/4374#issuecomment-454381465.

* Add test that dtype attribute still works if computed